requirements.txt
numpy~=1.26.4
matplotlib~=3.9.0
PyQt5~=5.15.10
pillow~=10.4.0
colorednoise~=2.2.0